Make default config responsible for country codes
Previously was done internally in code but could be overwritten
by aliases with the same name. Since some country codes have
multiple time zones, it is not easy to pick one zone with pytz
alone. For this reason, overwritting with alias is common and
and it would be more consistent if aliases handled all country
codes by default.